requirements.txt
functions-framework==3.4.0
flask==2.3.2
google-cloud-language==2.11.0
